fix structures of shop components
