+++ frontend/src/components/Shop/shops.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import {Shop, ShopType} from "./shop";
+
+
+const Shops: React.FC<{shops: Array<ShopType>}> = (props) => {
+    const shops = !!props.shops ? props.shops : []
+    return (
+        <div id="shops">
+            {shops.map(shop => {return (<Shop {...shop} key={shop.shopId} />)})}
+        </div>
+    )
+}
+
+export default Shops;
